backend/pyproject.toml
[tool.poetry]
name = "fossil-scribe"
version = "0.1.0"
description = ""
authors = ["Sol-Raksha Raja <solraksha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gpt4all = "^2.7.0"
nomic = "^3.1.1"
fastapi = "^0.111.1"
websocket = "^0.2.1"
llm = "^0.15"
gevent-websocket = "^0.10.1"
openai = "^1.37.1"
pyyaml = "^6.0.2"


[tool.poetry.group.test.dependencies]
pylint = "^3.2.6"
black = "^24.4.2"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
